+++ src/main/java/com/skillnest/everythingsouvneirs/EverythingSouvenirsApplication.java
@@ -8,17 +8,18 @@ import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.context.annotation.Bean;
 
 @SpringBootApplication
-public class EverythingSouvneirsApplication {
+public class EverythingSouvenirsApplication {
 
     public static void main(String[] args) {
-        SpringApplication.run(EverythingSouvneirsApplication.class, args);
+        SpringApplication.run(EverythingSouvenirsApplication.class, args);
     }
 
     @Bean
     public CommandLineRunner initUsers(UserService userService) {
         return args -> {
-            if (!userService.findByEmail("admin@everythingsouvenirs.ng").isPresent()) {
+            if (userService.findByEmail("admin@everythingsouvenirs.ng").isEmpty()) {
                 userService.createUser(new User(
+                        "123e4567-e89b-12d3-a456-426614174000",
                         "admin@everythingsouvenirs.ng",
                         "admin123",
                         "ADMIN",
