+++ src/main/java/com/skillnest/everythingsouvneirs/data/model/User.java
@@ -1,13 +1,21 @@
 package com.skillnest.everythingsouvneirs.data.model;
 
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
 import org.springframework.data.annotation.Id;
 import org.springframework.data.mongodb.core.mapping.Document;
+
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.Size;
 
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
 @Document(collection = "users")
 public class User {
+
     @Id
     private String id;
 
@@ -20,60 +28,9 @@ public class User {
     private String password;
 
     @NotBlank(message = "Role is required")
-    private String role; // e.g., "ADMIN", "USER"
+    private String role;
 
     @NotBlank(message = "Name is required")
     @Size(min = 2, max = 100, message = "Name must be between 2 and 100 characters")
     private String name;
-
-    // Constructors
-    public User() {}
-
-    public User(String email, String password, String role, String name) {
-        this.email = email;
-        this.password = password;
-        this.role = role;
-        this.name = name;
-    }
-
-    // Getters and Setters
-    public String getId() {
-        return id;
-    }
-
-    public void setId(String id) {
-        this.id = id;
-    }
-
-    public String getEmail() {
-        return email;
-    }
-
-    public void setEmail(String email) {
-        this.email = email;
-    }
-
-    public String getPassword() {
-        return password;
-    }
-
-    public void setPassword(String password) {
-        this.password = password;
-    }
-
-    public String getRole() {
-        return role;
-    }
-
-    public void setRole(String role) {
-        this.role = role;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
 }
