+++ src/main/java/com/skillnest/everythingsouvneirs/dtos/request/CreateQuoteRequest.java
@@ -1,29 +1,14 @@
-package com.skillnest.everythingsouvneirs.data.model;
+package com.skillnest.everythingsouvneirs.dtos.request;
 
 import com.skillnest.everythingsouvneirs.data.enums.QuoteStatus;
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.Pattern;
 import jakarta.validation.constraints.Size;
-import lombok.AllArgsConstructor;
 import lombok.Data;
-import lombok.NoArgsConstructor;
-import org.springframework.data.annotation.CreatedDate;
-import org.springframework.data.annotation.Id;
-import org.springframework.data.annotation.LastModifiedDate;
-import org.springframework.data.mongodb.core.mapping.Document;
-
-import java.time.LocalDateTime;
 
 @Data
-@NoArgsConstructor
-@AllArgsConstructor
-@Document(collection = "quote_requests")
-public class QuoteRequest {
-
-    @Id
-    private String id;
-
+public class CreateQuoteRequest {
     @NotBlank(message = "Name is required")
     @Size(min = 2, max = 100, message = "Name must be between 2 and 100 characters")
     private String name;
@@ -47,10 +32,4 @@ public class QuoteRequest {
     private String message;
 
     private QuoteStatus status = QuoteStatus.PENDING;
-
-    @CreatedDate
-    private LocalDateTime createdAt;
-
-    @LastModifiedDate
-    private LocalDateTime updatedAt;
 }
