+++ src/main/resources/application-dev.properties
@@ -1,20 +1,10 @@
 spring.application.name=Everything-Souvneirs
-
-# MongoDB Configuration
-spring.data.mongodb.database=everything-souvneirs
-spring.data.mongodb.uri=mongodb://localhost:27017/everything-souvneirs
-
-# JPA Removed - Not needed for MongoDB
-
+spring.data.mongodb.database=everything-souvenirs
+spring.data.mongodb.uri=mongodb+srv://oladimejivictor611:NYSh0JeKocC4Tmkl@everything-souvneirs.sijhpkl.mongodb.net/?retryWrites=true&w=majority&appName=Everything-Souvneirs
 server.port=8081
-#jwt.secret=your-secure-secret-key-256-bit-long-enough-for-hs256-algorithm
-#jwt.expiration=86400000
-
-# JWT
 #RESET_JWT_SECRET=Lro7KxAF6IvxTbQqynMFdzqHCL1m55CvaZ/V2M2ulawETobWLXXHZp3P5nMC+wMf401
 #RESET_JWT_EXPIRATION=120000
 
-# Cloudinary
 CLOUDINARY_CLOUD_NAME=dne74lnhk
 CLOUDINARY_API_KEY=919725957683897
 CLOUDINARY_API_SECRET=ikFY_3mXaiwBUeFMMZoD-k6Yljo
