+++ src/test/java/com/skillnest/everythingsouvneirs/EverythingSouvenirsApplicationTests.java
@@ -4,7 +4,7 @@ import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
 @SpringBootTest
-class EverythingSouvneirsApplicationTests {
+class EverythingSouvenirsApplicationTests {
 
     @Test
     void contextLoads() {
